restructure repo for easier navigation
