+++ api/adapter/adapter.proto
