+++ api/shim/shim.proto
@@ -18,6 +18,7 @@ message ClusterRequest {
 
 message ClusterResponse {
   string message = 1;
+  string snapshot = 2;
 }
 
 service Shim {
