+++ examples/grpc-godog-test/compliment_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/cucumber/godog"
 	"google.golang.org/grpc"
 
-	pb "github.com/zachmandeville/tester-prototype/example/complimenter"
+	pb "github.com/zachmandeville/tester-prototype/examples/grpc-godog-test/complimenter"
 )
 
 const (
