+++ examples/test-target/bootstrap-xds.yaml
