+++ examples/test-target/internal/resources/cache.go
