+++ examples/test-target/internal/resources/resource.go
