+++ tests/valid_target_test.go
